feat: Hoàn thiện API Đăng nhập, Đăng ký và module User/Role

+++ src/main/java/com/parking/management/module/building/BuildingRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BuildingRepository extends JpaRepository<Building, Long> {
+public interface BuildingRepository extends JpaRepository<Building, Integer> {
     // Custom query 1
     // List<Building> findBySomeField(String field);
 }
