+++ src/main/java/com/parking/management/module/vehicle/VehicleTypeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface VehicleTypeRepository extends JpaRepository<VehicleType, Long> {
+public interface VehicleTypeRepository extends JpaRepository<VehicleType, Integer> {
     // Custom query 1
     // List<VehicleType> findBySomeField(String field);
 }
